+++ pkg/providers/clientlists/data_akamai_clientlist_lists.go
@@ -292,5 +292,6 @@ func getValidListTypes() []string {
 		string(clientlists.TLSFingerprint),
 		string(clientlists.FileHash),
 		string(clientlists.USER),
+		string(clientlists.DOMAIN),
 	}
 }